update seed process to include API, shell-script and conn. string for access from outside container
