+++ server/app/db/schemas/database_schema.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class SeedDatabaseResponse(BaseModel):
+    message: str
+
