+++ server/app/restoreDataSeed_shell.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+echo "Running database restoreDataSeed script inside the container..."
+
+cd /server || exit
+
+python -m app.restoreDataSeed
+
+echo "Database seeding process completed!"
