+++ server/app/seed_shell.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+echo "Running database seed script inside the container..."
+
+cd /server || exit
+
+python -m app.seed
+
+echo "Database seeding process completed!"
